Start script boot integration
